Symbolic links automathically placed, if needed

+++ grow_for_pele.py
@@ -303,11 +303,14 @@ def main(complex_pdb, fragment_pdb, core_atom, fragment_atom, iterations, criter
         # Make a copy of the template file in growing_templates folder
         shutil.copy(os.path.join(path_to_templates, template_final), template)
 
-        # Running PELE simulation
+        # Creating results folder
         Helpers.folder_handler.check_and_create_results_folder(result)
         # Create symbolic links
-        os.symlink(c.PATH_TO_DATA, "Data")
-        os.symlink(c.PATH_TO_DATA, "Documents")
+        if not os.path.exists("Data"):
+            os.symlink(c.PATH_TO_DATA, "Data")
+        if not os.path.exists("Documents"):
+            os.symlink(c.PATH_TO_DATA, "Documents")
+        # Start simulation
         Growing.simulations_linker.simulation_runner(pele_dir, contrl, int(cpus))
         logger.info(c.LINES_MESSAGE)
         logger.info(c.FINISH_SIM_MESSAGE.format(result))
